+++ src/main/java/project/volunteer/DummyConfig.java
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Profile;
 import project.volunteer.domain.image.dao.ImageRepository;
 import project.volunteer.domain.participation.dao.ParticipantRepository;
 import project.volunteer.domain.recruitment.dao.RecruitmentRepository;
-import project.volunteer.domain.repeatPeriod.dao.RepeatPeriodRepository;
+import project.volunteer.domain.recruitment.dao.RepeatPeriodRepository;
 import project.volunteer.domain.scheduleParticipation.dao.ScheduleParticipationRepository;
 import project.volunteer.domain.sehedule.dao.ScheduleRepository;
 import project.volunteer.domain.image.dao.StorageRepository;
