+++ src/main/java/project/volunteer/domain/recruitment/application/RepeatPeriodService.java
@@ -1,0 +1,9 @@
+package project.volunteer.domain.recruitment.application;
+
+import project.volunteer.domain.recruitment.application.dto.RepeatPeriodParam;
+
+public interface RepeatPeriodService {
+
+    public void addRepeatPeriod(Long recruitmentNo, RepeatPeriodParam saveDto);
+
+}
