+++ src/main/java/project/volunteer/domain/recruitment/application/dto/RepeatPeriodParam.java
@@ -1,12 +1,12 @@
-package project.volunteer.domain.repeatPeriod.application.dto;
+package project.volunteer.domain.recruitment.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import project.volunteer.domain.repeatPeriod.domain.Day;
-import project.volunteer.domain.repeatPeriod.domain.Period;
-import project.volunteer.domain.repeatPeriod.domain.Week;
+import project.volunteer.domain.recruitment.domain.Day;
+import project.volunteer.domain.recruitment.domain.Period;
+import project.volunteer.domain.recruitment.domain.Week;
 
 import java.util.ArrayList;
 import java.util.List;
