+++ src/main/java/project/volunteer/domain/recruitment/domain/Week.java
@@ -1,4 +1,4 @@
-package project.volunteer.domain.repeatPeriod.domain;
+package project.volunteer.domain.recruitment.domain;
 
 import project.volunteer.global.common.converter.CodeCommonType;
 import project.volunteer.global.error.exception.BusinessException;
@@ -6,23 +6,23 @@ import project.volunteer.global.error.exception.ErrorCode;
 
 import java.util.Arrays;
 
-public enum Day implements CodeCommonType {
+public enum Week implements CodeCommonType {
 
-    MON(1, "월요일"), TUES(2, "화요일"), WED(3, "수요일"), THRU(4, "목요일"),
-    FRI(5,"금요일"), SAT(6,"토요일"), SUN(7,"일요일");
+    FIRST(1,"첫째주"), SECOND(2,"둘째주"), THIRD(3, "셋째주"), FOUR(4, "넷째주"),
+    FIVE(5,"다섯째주"), SIX(6,"여섯째주");
 
     private final Integer value;
     private final String viewName;
-    Day(Integer value, String label) {
+    Week(int value, String viewName) {
         this.value = value;
-        this.viewName = label;
+        this.viewName = viewName;
     }
 
-    public static Day of(String value){
-        return Arrays.stream(Day.values())
+    public static Week of(String value){
+        return Arrays.stream(Week.values())
                 .filter(v -> v.getId().equals(value.toUpperCase()))
                 .findAny()
-                .orElseThrow(() -> new BusinessException(ErrorCode.UNKNOWN_ENUM_VALUE, String.format("Day Value = [%d]", value)));
+                .orElseThrow(() -> new BusinessException(ErrorCode.UNKNOWN_ENUM_VALUE, String.format("Week Value = [%d]", value)));
     }
 
     @Override
@@ -33,7 +33,7 @@ public enum Day implements CodeCommonType {
     public String getDesc() {
         return this.viewName;
     }
-    public Integer getValue(){
+    public Integer getValue() {
         return this.value;
     }
 }
