--- src/main/java/project/volunteer/domain/repeatPeriod/application/RepeatPeriodService.java
@@ -1,9 +1,0 @@
-package project.volunteer.domain.repeatPeriod.application;
-
-import project.volunteer.domain.repeatPeriod.application.dto.RepeatPeriodParam;
-
-public interface RepeatPeriodService {
-
-    public void addRepeatPeriod(Long recruitmentNo, RepeatPeriodParam saveDto);
-
-}
