+++ src/test/java/project/volunteer/domain/recruitment/api/RecruitmentWriteControllerTest.java
@@ -19,12 +19,12 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
-import project.volunteer.domain.repeatPeriod.domain.Period;
+import project.volunteer.domain.recruitment.domain.Period;
 import project.volunteer.domain.image.dao.StorageRepository;
 import project.volunteer.domain.image.domain.Storage;
 import project.volunteer.domain.recruitment.domain.VolunteeringType;
-import project.volunteer.domain.repeatPeriod.domain.Day;
-import project.volunteer.domain.repeatPeriod.domain.Week;
+import project.volunteer.domain.recruitment.domain.Day;
+import project.volunteer.domain.recruitment.domain.Week;
 import project.volunteer.domain.user.dao.UserRepository;
 import project.volunteer.domain.user.domain.Gender;
 import project.volunteer.domain.user.domain.Role;
